web-app/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.31.0
torch==2.6.0
transformers==4.48.2
diffusers==0.32.2
Pillow==11.1.0
python-multipart
accelerate
pydantic
bcrypt
python-jose
huggingface_hub
